refactor(strategy): move MultiStrategyConfig and StrategyKind to hoprd
MultiStrategyConfig and StrategyKind are application-level config that
belongs in hoprd, not the library crate. Moving them there:
- Removes the only reference from hopr-strategy back to its own enum
- hopr-strategy now exports only Strategy, MultiStrategy, and per-strategy
  builders/configs — no application dispatch logic
- hoprd/strategy.rs owns StrategyKind, MultiStrategyConfig, build_strategies,
  and hopr_default_strategies

Also rename Cargo features for clarity:
  auto-funding       → strategy-auto-funding
  auto-redeeming     → strategy-auto-redeeming
  closure-finalizer  → strategy-closure-finalizer

Remove module-level doc comment from hoprd/strategy.rs per review suggestion.
Remove strum and serde_json from hopr-strategy (no longer needed there).
