+++ hoprd/hoprd/Cargo.toml
@@ -24,9 +24,9 @@ runtime-tokio = [
   "opentelemetry_sdk?/experimental_logs_batch_log_processor_with_async_runtime",
   "opentelemetry_sdk?/experimental_metrics_periodicreader_with_async_runtime",
   "hopr-strategy/runtime-tokio",
-  "hopr-strategy/auto-funding",
-  "hopr-strategy/auto-redeeming",
-  "hopr-strategy/closure-finalizer",
+  "hopr-strategy/strategy-auto-funding",
+  "hopr-strategy/strategy-auto-redeeming",
+  "hopr-strategy/strategy-closure-finalizer",
   "hopr-transport-p2p/runtime-tokio",
 ]
 telemetry = [
