+++ impls/strategy/Cargo.toml
@@ -17,9 +17,9 @@ default = []
 runtime-tokio = ["hopr-lib/runtime-tokio", "hopr-async-runtime/runtime-tokio"]
 telemetry = ["dep:hopr-metrics"]
 ## Strategy features — each enables one strategy implementation.
-auto-funding = []
-auto-redeeming = []
-closure-finalizer = []
+strategy-auto-funding = []
+strategy-auto-redeeming = []
+strategy-closure-finalizer = []
 
 [dependencies]
 anyhow = { workspace = true }
@@ -33,7 +33,6 @@ lazy_static = { workspace = true }
 serde = { workspace = true }
 serde_with = { workspace = true }
 smart-default = { workspace = true }
-strum = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
 validator = { workspace = true }
@@ -45,7 +44,6 @@ moka = { workspace = true, features = ["sync"] }
 
 [dev-dependencies]
 chrono = { workspace = true }
-serde_json = { workspace = true }
 hex-literal = { workspace = true }
 hopr-lib = { workspace = true, features = ["runtime-tokio"] }
 hopr-async-runtime = { workspace = true, features = ["runtime-tokio"] }
